Various enhancements
Removes the need for random callback names by using __return as the
callback which is unlikely to be used in regular code.

Adds async support if statements that have return in them, also adds
a return after each callback to prevent more than one callback being
called.

Need to add support for other control structures like for, while etc...

Makes sure root level has a return callback even if one is set inside
an if statement incase it is not evaluated all functions should call
the callback and return.

Better support for assigning variables to await functions, supports
binop assignments for variables that are already declared and correctly
assigns the variable in the right scope.

Added existing library continuation to see how it works and how they
handle certain situations.