Error throwing in async functions
Support for throw statement in async functions which is replaced with
a callback using the "error first" standard with the error thrown
as the paramater.

Added an unreachable callback error when a return is used after another
return or throw statement and could not be reached in real world usage.

Fixed all the positiosn to use the correct positions form the
expressions to report errors properly.