Various enhancements
Don't replace "throw" with error callback if inside a try block

Add if, meta, try for await macro.

Handling of if statements, adds __if callback to the top of if block
and moves the currentBlock into that callback to handle. The __if
callback is only added if an async call is made within otherwise it is
ignored.