expr & call stack
Decided not to move blocks inside the while loop and follow the same
pattern as the if block.

Added expr stack and call stack to track the nested path of exprs, only
adding important ones like for, while, if that require information about
parent exprs.

Not 100% sure how this will work yet but need some way to handle nested
expressions, current use of "isInIf" & "isCalled etc do not work well
and cause all nested expressions to assume "isCalled" when they might
not need to use an async expression like "if" inside an async while.

expr stack should give enough information to use the correct method
depending on the situation.