Simple expr stack
Simpified the expr stack concept, using the expr stack to check for
nested if etc... we can easily work out what kind of calls have been
made.

Current problem is adding to the stack after the current depth because
of the current method of adding to the stack... considering creating a
new class and object for expressions so the information can bestored
in the expr instance instead of on the class which is the current
problem.