Testing methods for determinig expression states
Current expr stack method works well enough to figure out the expr path
but does not work for pushing after the fact so cannot be used to
tell if a callback was used.

Testing method of giving each expression a unique id and then pushing
the unique id of the last expression on the current stack, can test
for called back by comparing the call stack with the expression stack
ids which if they overlap indicates a callback within the expression
stack.