Fix bug with if statements
If no else block is supplied for an if statement we now add an empty
block so the after if callback is always called and the code after
the if block is reached.

Changed what is added to the callstack so a deeply nested async call
correctly transforms parent if blocks.