Fixes do/while control expressions and others
findExprs function will fetch a named expression or expressions
recursively, can supply an ignore list of expresions to skip.
Only really works for break/continue at the moment and ignore is used
to skip other for and while loops. This is used to hunt down locations
of continue/break so we can replace them later if a async call was used.

Fixed continue for do-while block by adding the condition as well.

Removed checks for last return since it doesn't really matter.