Adding extra data to the "expr stack"
Adding more data to the expr stack to replace some of the object vars
that are storing information like "currentMetadata" and "currentBinop".

Instead of storing these in variables which have the potential to be
overwritten we can just search the stack for the information we need.

The only one that falls outside of this is "currentVar" because it is
not an expr we cannot push it to the stack, we might have to do
something similar to "ElseIf" and "Else".