ExtJS macro
Converted the extjs experiement for extjs class support to a macro,
adding the extjs build macro will add functions to override exts native
extend functionality in favour of haxes extend, it also copies the
config and mixins property and applies them to the extension config
so preprocessors can be run on them. This lets us support mixins but
the mixin functions will still have to be defined as externs in order
to access them.